refactor: Refactored what was reported by coderabbit, reviewer and sonarCloud
Summary:

- Renamed the modules to follow the reviewers standard.
