+++ devdox/tests/unit_test/test_doubles/app/repository/user_repository_doubles.py
@@ -2,7 +2,7 @@ import uuid
 
 from models import User
 
-from app.repositories.user_repository import IUserStore
+from app.repositories.user import IUserStore
 
 
 class FakeUserStore(IUserStore):
